[Gold V] Title: 다각형의 면적, Time: 0 ms, Memory: 2100 KB -BaekjoonHub

+++ 백준/Gold/2166. 다각형의 면적/다각형의 면적.cc
@@ -3,7 +3,7 @@
 using namespace std;
 
 int N;
-pair<double, double> point[10000];
+pair<int, int> point[10000];
 
 double CCW(pair<double, double> p1, pair<double, double> p2, pair<double, double> p3) {
     double res;
